requirements.txt
flask==2.3.3
pandas==2.1.1
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0